Implement new gateway directoryD service
Summary:
This diff adds a new RPC service to the DirectoryD proto
and implements it. This service allows for directory records containing
a location history as well as other arbitrary identifiers to be
stored locally in Redis. The gateway state replication service
will then replicate these back to the Orchestrator, which can then be
queried from the API/NMS.

Reviewed By: xjtian

Differential Revision: D18239176

fbshipit-source-id: b098ab6a3bfd9964faf65892827dca52d82b39b4