Move gateway directoryD service from LTE module to Orc8r
Summary:
The DirectoryD service is a generic gateway service that allows for
storing record to have those replicated back to the Orchestrator. Additionally,
the proto is already defined in the Orchestrator module. As such, this service
should reside in the orc8r/gateway as opposed to lte/gateway. This diff implements
that change.

Reviewed By: xjtian

Differential Revision: D18243453

fbshipit-source-id: e8fc5c66a5e9c8ba915912847cc3fa074f754a5f