create ipv6 ping engine (#841)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/841

This diff creates an IPV6 ping engine. The ping engine opens an ipv6 socket. It doesn't yet send or read them (that's in the next diff of the stack).

Docker for mac doesn't work with ipv6 but this diff handles a failed socket open gracefully. Whether or not a machine supports v6, this won't break it.

Reviewed By: Mokon

Differential Revision: D18434701

fbshipit-source-id: bc37d32e86d275671a7a81a1c2ecc6856876455e