Update orchestrator's directoryD service to fetch from state service (#865)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/865

This diff updates the Orchestrator's directoryD service
to fetch from the state service for IMSI_TO_HWID's. It also updates
the feg_relay service to properly fetch using a networkID since the state
service keys on (networkID, deviceID, type). For now we will
just iterate through all servedIDs to find the directory state.

In the future, if we ever need to support a large number of served_ids
for a given FeG network, this can be optimized by using a secondary index that maps
IMSI (id) to networkID.

Reviewed By: xjtian

Differential Revision: D18368687

fbshipit-source-id: c1cb4549e5f0fb20b7e29ac41a1387d5a3241ac7