Only display graph legend with few enough labels
Summary: If a graph had a bunch of metric series, the labels would squash the graph, making it unusable. I couldn't figure out how to make the legend a fixed size and scrollable, so this stopgap fix just removes the legend if it has more than 5 labels. In order to maintain usability I fixed the tooltip that displays when you hover over datapoints. It now shows the labels at that point. Additionally I made the hover hit size bigger so it's a lot easier to hit the point with your mouse.

Reviewed By: murtadha

Differential Revision: D18530839

fbshipit-source-id: 79fc329cf040322e706cb05ee714992eaaa9e099