longer snmp timeout
Summary:
Increase SNMP timeout to 50 seconds (instead of 5 seconds) and polling interval to 55 seconds instead of 10 seconds.

I increased the SNMP timeout because we were seeing issues with ENS's core switches. They have lots of interfaces and as such the many snmp requests get slowed down a LOT. I created a task to batch our SNMP requests, which will hopefully quicken everything, and we can lower the timeout to a reasonable length.

By making the polling interval 55s, since the UI updates every minute, we should still see an update.

This slows the tests down a bit (the ones that wait for snmp timeouts). We'll address that when this is resolved.

Reviewed By: Mokon

Differential Revision: D18620310

fbshipit-source-id: ab4de994408ac6d5226d182048fec88c42e9c12f