Modify UseGyForAuthOnly case to only send the bare minimum CCR-I
Summary:
When we are using USE_GY_FOR_AUTH_ONLY, we do not care about any rating group that might be returned as part of gx CCA-I. We will always send a gy CCR-I with no MSCC specified.
To simplify the logic, I've created a separate function that deals with this specific case that CreateSession will call.

Differential Revision: D18648956

fbshipit-source-id: c3faeeec5ca3c642a4ba8221f4d2cd69ca64904a