Modify UseGyForAuthOnly case to only send the bare minimum CCR-I (#928)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/928

When we are using USE_GY_FOR_AUTH_ONLY, we do not care about any rating group that might be returned as part of gx CCA-I. We will always send a gy CCR-I with no MSCC specified. My previous impression was that we send a CCR-I to gy ~regardless~ of whether there is a rating group or not.
To simplify the logic, I've created a separate function that deals with this specific case that CreateSession will call.

Reviewed By: emakeev

Differential Revision: D18648956

fbshipit-source-id: 13056b8db99a4448cd5e666a95e911d653fa2dc2