Update diameter services to maintain connections when active (#938)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/938

When the health manager on the FeG sends ENABLE to diameter services,
the current behavior is to ensure connection creation is enabled. However, a
connection is not created if one does not exist. This means that after a
failover, the Active FeG services won't have connection to core services.
This causes errors on the core side. Additionally, messages that originate
from the core (RTR, RAR, etc.) won't be handled properly until a connection
is created from a AG -> FeG request. This diff fixes this by ensuring a connection
exists (GetConnection) on each Enable call.

Reviewed By: themarwhal

Differential Revision: D18627673

fbshipit-source-id: d5ed050d78254da33cb8f3cc0de8f505451ce45a