Get interface indices via snmpwalk (#927)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/927

The getNumberOfInterfaces function returns the number of physical and virtual interfaces, wich messes with our logic of reading interface indices. This gets the interface indices by doing an snmpwalk over the mib which has those interfaces.

This is slightly less efficient because walk happens in series instead of parallel, but the next diff in this stack will save the indices so that they're not read repeatedly retrieved, leading to an overall perf gain.

Reviewed By: Mokon

Differential Revision: D18645543

fbshipit-source-id: 031bb783450325604ef235938c24025cd4a63cb0