Add session id from sessiond to subscriber's directory record
Summary:
This diff makes a call from SesisonD to DirectoryD
to update a subscriber record with the current session_id.
By storing the session_id in Redis, we are able to recover from a sessiond
restart by loading all previous session_ids and sending CCR-Ts for each.
This functionality will be implemented in a stacked diff.

Differential Revision: D18702429

fbshipit-source-id: c26c63aea966ef1eb765c16d22e24cc936524e41