Add session id from sessiond to subscriber's directory record
Summary:
This diff makes a call from SesisonD to DirectoryD
to update a subscriber record with the current session_id.
By storing the session_id in Redis, we are able to recover from a sessiond
restart by loading all previous session_ids and sending CCR-Ts for each.
This functionality will be implemented in a stacked diff.

Reviewed By: themarwhal

Differential Revision: D18702429

fbshipit-source-id: fcb8740bb962aa72007f472629b4c806526e5dfd