Standalone PCRF installs static policy rules registered to subscribers
Summary:
Captive portal functionality is being made less of a special case from the rest of local PCRF/OCS.

For all static policy rules assigned to users, including redirection for captive portal, it is suggested that network operators use the NMS to register the policies to the subscribers. From there, policydb will act as the local PCRF and get pipelined to install the rules for the subscriber.

Captive portal will now be installed per user and configured the same as any other rule on the local PCRF/OCS

Reviewed By: xjtian

Differential Revision: D18737090

fbshipit-source-id: 7f2cdfa544156f8b2c15622c41809b1421144ac8