Add `use_local_session_proxy` to unbreak sessiond <-> session proxy connection in the CWAG integration test (#979)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/979

The CWAG Integration test relies on talking to a locally running Session Proxy service to run the test. Originally, it used a flag in sessiond to specifiy a local port. But sessiond was modified in D18073209 to no longer use these flags. Leading to a http `502` error when sessiond tries to talk to sessionProxy.
I'm putting back in a similar piece of code back into sessiond, but to avoid confusion I'm now calling this flag `use_local_session_proxy` and documented that it is used for testing.

Differential Revision: D18837829

fbshipit-source-id: 9d152e15a05cedadf9bd4772d0e1b5e30c51944a