Fix CI integ test
Summary:
In the current setup of the integration test, we do the following:
1. setup cwag test networking in the test VM
2. start the services in the magma VM
3. start the tests in the test VM

Somehow, the route added in the first step would be removed when step 3 happens. I'm not 100% sure on the root cause but my hunch is that we run the ansible script when we try to ssh into the test VM again.

I'll probably change the setup so that we don't rerun the ansible script again in the future

Reviewed By: koolzz

Differential Revision: D18895265

fbshipit-source-id: 4661d15d0c59e263b2356a5e74f765b51a949c38