Support reception of empty HTTP request after provisioning session end
Summary:
After a TR-069 provisioning session, which is ended by `enodebd` sending an empty HTTP response, it is expected that the eNodeB will not send any TR-069 messages other than an `Inform` to start a new provisioning session. It is found that for at least one device/software combination, the eNodeB will send an empty HTTP request, which is unhandled and will cause `enodebd` to go into an error state.

## Changes
Service `enodebd` will handle an empty HTTP request outside of a TR-069 provisioning session by ignoring it.

Differential Revision: D18901078

fbshipit-source-id: 9aee826063b936c09189b0ce0783f351982e0f96