Support reception of empty HTTP request after provisioning session end (#993)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/993

After a TR-069 provisioning session, which is ended by `enodebd` sending an empty HTTP response, it is expected that the eNodeB will not send any TR-069 messages other than an `Inform` to start a new provisioning session. It is found that for at least one device/software combination, the eNodeB will send an empty HTTP request, which is unhandled and will cause `enodebd` to go into an error state.

## Changes
Service `enodebd` will handle an empty HTTP request outside of a TR-069 provisioning session by ignoring it.

Reviewed By: xjtian

Differential Revision: D18901078

fbshipit-source-id: 743c4c3545d6da7eed34a05cd69272689ada5e52