PromQL Parser
Summary:
Implements a parser for a strict subset of PromQL. This parser takes an input PromQL string and returns a javascript object that implements the `PromQL.Expression` interface.
* The parser is generated using the nearley package, from the grammar defined in `PromQLGrammar.ne`
  * To generate the parser, run `nearleyc fbcnms-packages/fbcnms-alarms/components/prometheus/PromQLGrammer.ne -o fbcnms-packages/fbcnms-alarms/components/prometheus/__generated__/PromQLGrammar.js`. You need the nearley node package: `npm install -g nearley`
* I have not implemented a few features of the language, e.g. the `by` clause of `sum(metric) by (label)`
  * I will add this in a stacked diff to show how to iterate on the parser by just modifying the grammar.

Javascript Note: I have disabled eslint on the generated PromQLGrammar.js file since it generates it in a way that causes a bunch of lint errors. I believe the tests should be sufficient to feel confident that this works without issue.

Note: I made a change to the tokenizer where I parse the `PromQL.Range` object in the tokenizer. This is the best solution to the problem of how to get the range values out of the string in my opinion. Since the tokenizer matches a regex, you can parse the string without having to worry about it being in a different format than you expect.

Reviewed By: aclave1

Differential Revision: D18897793

fbshipit-source-id: 478ae9294b55b579321526df508ffb2a8b254ea9