Add support for tombstoning objects in Redis
Summary:
This diff adds the ability to mark an object as garbage
is redis, without directly deleting the object itself. This functionality
allows for a garbage collector to iterate over all garbage and delete those
states from the Orchestrator.

The implementation adds to the RedisState proto (redis object metadata)
an `is_garbage` field. The RedisFlatDict is updated to treat garbage objects
as deleted, such that clients see no functional difference between a deleted
object and a garbage object.

While using metadata stored with the object leads to a less efficient implementation,
it is favored over an alternative implementation of deleting the object and storing a
list of garbage type:keys as it allows us to still read the garbage collected data,
which may be needed for global resource allocation.

Differential Revision: D18840294

fbshipit-source-id: 7c5fdb5e7c81a09cefd927094bf55b7cce66f89c