Implement garbage collector for state service
Summary:
This diff implements garbage collection functionality for
the state service. This will ensure that "deleted" state on the gateway
is propagated back to the Orchestrator.

The garbage collector works by iterating through all garbage keys
of each configured redis dict, and then sending a single DeleteStates RPC
call to the Orchestrator. If this call succeeds, the state is
deleted locally from Redis.

*One key note*: State replication and garbage collection do not run at the
same time. Currently garbage collection is run after 5 iterations of replication.
This is purposedly done to help avoid race conditions for State in the Orchestrator
without having to lock every key for a long amount of time (read, RPC call and potential delete).

To illustrate this, consider:
    t1 (replication): Load all keys (N is large) of non-garbage state
    t1 (replication): Read redis key v and add it to replication request
    t3 (application): Marks redis key v as garbage
    t2 (gc): Begin running iteration, loading all keys (N is small) of garbage state
    t2 (gc): Send DeleteStates RPC call
    t1 (replication): Finishes aggregating request, sending RPC call to replicate state for key v

The result is state that shouldn't exist in the Orchestrator, without that state existing on the gateway.

Reviewed By: xjtian

Differential Revision: D18845828

fbshipit-source-id: 585ad51dc9d06d79944d0dff54e76ac0f9d69eee