Add cleanup functionality to state service for keys that have been deleted
Summary:
This diff corrects a bug that currently exists in the state service where if a
key gets deleted and then re-added to Redis, the addition may not be replicated to
the state service if the key has the same version as it had before it was deleted.

This diff fixes this issue by keeping track of which keys have been seen each iteration.
After replicating state, the state replicator cleans up it's internal dict to only store
this keys. Thus any deleted keys, will be removed from this internal dictionary.

Reviewed By: andreilee

Differential Revision: D18837820

fbshipit-source-id: a611c2abeee987d06be8eee652a19ba4aee068e5