PortDefinitionsAddEditTable import typo (#137)
Summary:
Pull Request resolved: https://github.com/facebookincubator/symphony/pull/137

Similar to previous diff on stack, the import was incorrect, and the lint was ignored/missed.
In this case, we were missing the curly brackets on import.

This also fixes the fact that the fragment wanted an array, but wasn't specified. Add `relay(plural: true)` to the fragment graphql.
Rename the variable to be plural as a result.

Reviewed By: murtadha

Differential Revision: D18940845

fbshipit-source-id: 60b155c5d6ea887fa033d1d7096416412444d914