Handle redirection rules in pipelined setup
Summary:
For now just reinsert redirection rules, this is a bit of a hack but redirection relies on async dns request to be setup and we can't currently do this from out synchronous setup request. So just reinsert

Its not pretty but it'll do for now. The downside to this reinsertion is that it might interrupt the tcp connection and we'll lost pkt count for that rule

Reviewed By: amarpad

Differential Revision: D17696619

fbshipit-source-id: 999c3740f07c056e58463f4a14834955e84debe5