Making ovs flows persists across MME restart
Summary:
The MME process always cleans up all the OVS when the switch connects to the
controller. This change adds a check that flows are only removed when the
`persist_state` flag is false. Instead of triggering the flow removal through
BaseApplication, we can do it through the OpenflowController object itself.

Also fixed some formatting.

Reviewed By: koolzz

Differential Revision: D18981364

fbshipit-source-id: 7aa3b63247781c55eceef3214f8e09a6bd80f948