refactor device model in devmand to be multiple platforms
Summary:
This commit begins the refactor of the device model. It is mainly
moving some files around for better organization and renaming some classes.

I also get fix a few issues with the structured cli device and make it
configurable via the ui.

I added a non-used fscache service which will be extended in a future commit.

Differential Revision: D18818502

fbshipit-source-id: 47d5a235d74f06d8e8b910350ee5e6140698d3b7