Simplify report_termination calls (#1038)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1038

`report_termination` currently takes in a callback function that gets called after the reporting grpc call is made. Since we don't do anything other than log the status of the call, I added a separate function that returns a callback fn that logs the result, to clean up the codebase a bit.

Eventually, we should properly handle the error since there could be many reasons behind the error.

Reviewed By: xjtian

Differential Revision: D18943087

fbshipit-source-id: e45a89b63454d04b9d2d0ed9654cdc97ca652af1