Make ReloadPrometheus a configmanager client function
Summary:
This small refactor from having `reloadPrometheus` being a helper function in the handlers file to being a method of the PrometheusAlertClient makes the method calls cleaner since we can just pass the alertmanager URl to the client on construction. Also allows the return of `ReloadPrometheus` to be controlled and tested for in unit tests.
* Also add some unit tests
* Small refactorings of test file to clean it up

Reviewed By: xjtian

Differential Revision: D19134511

fbshipit-source-id: bd51c23dc847a13b082b9eb3cc06e6f38aac4f46