Make ReloadAlertmanager a configmanager client function
Summary:
This small refactor from having `reloadAlertmanager` being a helper function in the handlers file to being a method of the AlertmanagerClient makes the method calls cleaner since we can just pass the alertmanager URl to the client on construction. Also allows the return of `ReloadAlertmanager` to be controlled and tested for in unit tests.
* Also add some unit tests
* Small refactorings of test file to clean it up

Reviewed By: xjtian

Differential Revision: D19134503

fbshipit-source-id: 2a24c0f829a57968a3febffe5b8e1f3583df1a23