Make ReloadPrometheus a configmanager client function (#1043)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1043

This small refactor from having `reloadPrometheus` being a helper function in the handlers file to being a method of the PrometheusAlertClient makes the method calls cleaner since we can just pass the alertmanager URl to the client on construction. Also allows the return of `ReloadPrometheus` to be controlled and tested for in unit tests.
* Also add some unit tests
* Small refactorings of test file to clean it up

Reviewed By: xjtian

Differential Revision: D19134511

fbshipit-source-id: 8082f73ec5991afea98ab39fee55fdfc79da79f9