Add optimization to gateway GetUpdates logic allowing intelligent non-updates to mconfigs.
Summary:
GatewayConfigsMetadata now includes a GatewayConfigsDigest, storing the hash of a GatewayConfigs (hash taken with empty metadata field). The gateway passes this digest to GetUpdates via the extraArgs parameter. The cloud still generates the new GatewayConfigs for the requesting gateway, but compares its generated digest with the received digest; if the digests are equal, an empty UpdateBatch is returned to the gateway.

Originally the plan was to generate the digest both at the cloud and gateway, without storing any new information in GatewayConfigsMetadata. However, canonical bytewise representation is a non-feature for protobufs, hindering this approach. Workarounds, such as a JSON-based canonical representation, were considered, but ultimately deemed undesirably hacky.

Thus, since golang's protobuf implementation supports a deterministic bytewise encoding, the cloud generates the digest and passes it to the gateway via a new field in GatewayConfigsMetadata.

Reviewed By: xjtian

Differential Revision: D18918909

fbshipit-source-id: 671fecb796b6d4d5f0a8945865b8fa5674cdd6f6