Add support for non-ASCII rules' monitoring keys
Summary:
Add support for non-ASCII rules' monitoring keys.
Monitoring Keys are Octet strings and should be entered as such. We use base64 encoding for them (on swagger - REST API).
Any valid base64 encoded MonitoringKey will be decoded before PolicyRule is streamed as proto message.
If a MonitoringKey is NOT a valid base64 string, it'll be passed as is (string to []byte) to the streamer. So, monitoring key such as "cat" will be passed as is, but 'dGhpcyBpcyBhIHJ1bGU=' will be decoded.
Therefore it's advisable to always base64 encode MonitoringKeys for static rules during creation.

Reviewed By: themarwhal

Differential Revision: D19166632

fbshipit-source-id: f92c10eeff77e08803547c612af56770efe66f39