Fix crash on WorkOrder Project change
Summary:
Changing WorkOrder's Project will result in a crash since in WorkOrder's Header trying to use the Project's type for Breadcrumbs tooltip and the selected Project returned from the ProjectsTypeahead doesn't contain the full project type (but the name only).
Need to improve the fix so the selected Project will include the type object as well.

Reviewed By: noamsch

Differential Revision: D19211394

fbshipit-source-id: 623d6aecd5d3d9fc52e6829cad5a82fdddf7527b