Change HSS Lite to not to modify "read only" configs on Start (#1080)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1080

Change HSS Lite to not to modify "read only" configs on Start
HSS Lite needs to communicate back to the caller it's "real" local address when it's
started without preconfigured port. This kind of (port-less) configuration is used by
unit tests to avoid test failures due to port availability/conflicts.
Currently HSS Lite modifies its own config during blocking Start() call which is not
thread safe as well as non-obvious for the caller. This diff changes Start's completion
chan to be used to communicate local address as well as start completion.

Differential Revision: D19335959

fbshipit-source-id: e21ce262d70b468e80beb1b14ec8eef961a29948