Fix wizard of connect ports
Summary:
Had a bug where if you had a location filter in the equipment search, clicked next and then back, the wizard would crash.
This happened because the filter config for the existing value wasn't present.

Note: we probably need to refactor the search in overall, as it's pretty flakey and the flow types aren't the greatest.

Reviewed By: AmitArbel

Differential Revision: D19390967

fbshipit-source-id: 844164575ee7357a9fd690ca5b79933b18d8f272