fixing "you are not logged in" notice when user is still logged in (#246)
Summary:
Pull Request resolved: https://github.com/facebookincubator/symphony/pull/246

Some people are having issues with the login notice popping up in workorders even when they were logged in. Not sure why they're requesting routes they don't have permission for (could be a cookie issue?), but I took a look at the code again and it looks like I didn't actually check to see if the user was logged out before returning the 403 error.  This should fix the notice popping up when they're logged in.

Reviewed By: rckclmbr

Differential Revision: D19519916

fbshipit-source-id: 0c59dc2dcce59293134fb19ff638134d94fe7ea2