Add basic error handling for CCA-Update permanent failure (#1134)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1134

This diff adds some error handling for CCA-Updates for permanent diameter error codes.  (between 5000 & 6000).
For usage monitor updates and charging credit updates, if we receive a permanent error, we should terminate that service. Since SessionProxy sends each monitoring update and charging update as separate requests, this is error handling on a command level.
(It's not great that we are leaking session proxy logic into sessiond, but that has to be tackled with a much larger refactor)

Reviewed By: xjtian

Differential Revision: D19538531

fbshipit-source-id: a1f9445494827f1e6f7f0e8c6c9d35b45d6e7849