Add http_config support to alertmanager configurer (#1130)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1130

Adds `http_config` support to alertmanger configurer. This struct is used both globally for alertmanager and in some receiver configs.
http_config is how alertmanager is configured to speak to http services. This includes basic auth, proxy_url, etc. This applies to the global alertmanager configuration, as well as a few receiver types.
* Due to the way prometheus coded their structs, I had to copy some of them and make small modifications to enable JSON marshaling to work
* Had to rearrange some directories to avoid import cycles

There's no need for this functionality in orc8r so I didn't update the API to support it.

Reviewed By: xjtian

Differential Revision: D19623469

fbshipit-source-id: 8ab86074daa5dc545839662b007badc25eb07fc5