Using IMSI from ITTI for SGW/PGW messages (#1137)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1137

- Setting and retrieving IMSI on msg header from MME, pcef_handlers to be used by SGW/PGW itti messages
- Adding get_associated_imsi on SGW and PGW task to retrieve IMSI, adding debug messages to log retrieved IMSI

This will enable SGW and PGW tasks with a defined method to retrieve the IMSI, this will be used to retrieve only specific state per UE from redis, so the entire state shouldn't be retrieved and parsed, just for the UE that the messages are processing. Also this will help for a cleaner way to add IMSI prefix to all logging messages.

Reviewed By: koolzz

Differential Revision: D19627417

fbshipit-source-id: 6cf17abfeda2c7b746bef9b7be5fbc25b9deb940