symphony/exporter/links: fix bug that that caused erros/panic (#260)
Summary:
Pull Request resolved: https://github.com/facebookincubator/symphony/pull/260

when on the for loop for each csv line - I collected the errors to verify the file structure is valid.
On each error I used "continue" in order to skip to the next line.

I had a nested for (for the ports) and i used continue in it as well - which didn't exit the row-run.

Change - change it from continue to break and added a flag to verify the loop ended successfully.

Reviewed By: alexsn

Differential Revision: D19643068

fbshipit-source-id: c40ddc989dec67d24ca01bdc874eb16204a48a19