test normal users can only see their networks
Summary: I wrote this test in an attempt to reproduce an error a partner was experiencing.  It turns out there's no bug in the code and the partner was simply on a very old revision, but we might as well have this test for the future to make sure this doesn't break.

Reviewed By: rckclmbr

Differential Revision: D19672542

fbshipit-source-id: 15208ec1da260c16d8ff0b6ce409512357f2b745