Fix policies page when there is white space
Summary:
Space in the URL was breaking the request. This might be better handled by the API library, but fixing that requires a PR and an update to the library, so for now doing it this way.

If we think making the library handle this (which could also be potentially risky), I can add a TODO to revert this

Reviewed By: rckclmbr

Differential Revision: D19720191

fbshipit-source-id: 72da53fd24cd11f0cc4c6dfa3b4d27992a9cfabe