Add/remove helper functions to clean up `init_session_credit` (#1164)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1164

Making a helper function `handle_session_init_rule_updates` to move all of rule install/removal related logic into this one. Motivation is to make `init_session_credit` a bit easier on the eyes.

Move the logic of `get_rules_from_policy_reauth_request` into the parent function. Since the function doesn't do much, it's probably better to keep it in the parent function so that the logic is not as fragmented.

Reviewed By: ulaskozat

Differential Revision: D19750624

fbshipit-source-id: b5b2d9c2628c8a33f9f7c04325e76d520d614618