Add struct definitions for marshaling/unmarshaling of session state
Summary:
movethingspatrick

**For Stateless sessiond**

The idea here is to just help get sessiond state stored in Redis for persistence, instead of always being held in memory.

To do that, each class from `SessionState` and below need to be able to be marshaled/unmarshaled between a format appropriate for storage, either as serialized protobufs or serialized swagger models. As an intermediate step to this, each class will be responsible for a marshal/unmarshal method to change back and forth between data-only structs. From there, the storage interface, `SessionStore` will be responsible for getting the struct into storage.

Reviewed By: xjtian

Differential Revision: D19673085

fbshipit-source-id: 49cfd1f40024ac0f32ed2a10c48244b2f763f11d