Update CWAG pkg repo to use test repository
Summary:
We can't have different debian packages with the
same name for the different distributions all in the same debian
repository. As such, we've created a new repo 'test' for CWF.

This diff updates the ansible role for the CWAG to use this repo.
This should fix failing CI tests.

Reviewed By: themarwhal

Differential Revision: D19797729

fbshipit-source-id: d2f51d6a0e87f4eef046de30ff2671951a168cdc