Add events-based MemorySessionStore for handling session data access
Summary:
To enable stateless sessiond, a storage interface is required to have session states read and written on a per-transaction basis. As a first step towards this, the `SessionStore` class is being added to get the data for a session, make arbitrary writes on it, and then commit it to storage. `MemorySessionStore` is used to work out the interface here, though `RedisSessionStore` will be written to handle proper persistence.

Here, since gRPC calls are handled concurrently by sessiond, there is a need to allow only one call to have R/W access on a subscriber at a time, and the SessionStore handles these details.

 ---

## Changes
- Addition of `MemorySessionStore` to handle R/W of sessiond state, a non-persistent version of `SessionStore` which will be used to ensure that the interface works well for `session_manager`
- Added unit tests to verify that `MemorySessionStore` works as intended

 ---

Reviewed By: ardzoht

Differential Revision: D19783553

fbshipit-source-id: e3336f876f31e7dfe232e1900af64f8e168fdcae