symponh/ connect port wizard: remove duplicate wizard set
Summary:
removed wizard set on locationFilter
it caused re-rendering of the component as the props changes, and for some reason props came empty.

wizardContextSet happens on "onBlur" either way (and until we Blur there isn't a reason for us to set it to the context.

wdyt? did I miss some functionality that this code does?

Reviewed By: AmitArbel

Differential Revision: D19768048

fbshipit-source-id: 0ff44d08236f89a0f91e99839ed491024a521fef