Add get_number_of_monitoring_rules to RuleStore
Summary:
## Monitoring Quota Exhaustion Detection for CWF use case
When Quota exhaustion is detected in the PCRF, it will send rule removals for the existing rules. This means that the session is out of monitoring quota once there are no rules installed tied to a monitor.
This diff exposes a function `get_number_of_monitoring_rules` to sessiond's `RuleStore`. This way, at the end of each UpdateSession or ReAuthRequest, we can check if there are any existing monitoring rules in the system tied to the session.
Once there is none, we can propagate that information to pipelined.

This also includes a very minor refactoring in the session state test.

Reviewed By: xjtian

Differential Revision: D19728231

fbshipit-source-id: 2ebc0c6cd160377be939ad6998e5559f126cf4c5