Update integ test to reflect quota exhaustion change
Summary:
The behavior of cwag sessiond changed in the previous diff for quota exhaustion detection so reflecting that in the tests.

`auth_ul_test.go`: The default behavior when volume is not specified is to run traffic for 5 seconds. (default behavior for iperf3) Changing that to 100k bytes here.

`omni_rules_test.go`: Install a block-all static rule with monitoring. Override that with a pass all omni rule with a high priority.

Differential Revision: D19863837

fbshipit-source-id: 2d58265b4089bdad071161a66e765e9d00dfc60c