Update integ test to reflect quota exhaustion change
Summary:
The behavior of cwag sessiond changed in the previous diff for quota exhaustion detection so reflecting that in the tests.

`auth_ul_test.go`: The default behavior when volume is not specified is to run traffic for 5 seconds. (default behavior for iperf3) Changing that to 100k bytes here.

`omni_rules_test.go`: Install a block-all static rule with monitoring. Override that with a pass all omni rule with a high priority.

Reviewed By: mpgermano

Differential Revision: D19863837

fbshipit-source-id: 14d1dc1768792b0ab2c666f720c1c67ca3d45b0c