Terminate session on monitoring quota exhaustion on RAR/CCA-U (#1221)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1221

# Connection Steering Feature Summary
CWAG will host a flask server to indicate whether the subscriber has quota or not.

If a subscriber tries to connect without any quota, the subscriber will stay connected for 30 seconds (configurable) but without any access to data. During this time, curling the flask server  should indicate that the subscriber is out of quota.
Once the 30 seconds passes, the subscriber will get disconnected.

If a subscriber runs out of quota mid-session, the subscriber will be disconnected immediately.

# Implementation Details
Currently, the way we configure policies for the CWF case is that only WLAN rules are monitored and tracked by the PCRF. This means that whenever a CWF session no longer has monitoring rules, the subscriber is out of quota.

# What's changed in this diff
This diff implements the case where a subscriber runs out of quota mid-session. In this case, sessiond will receive a CCA-Update or a RAR with rule removals for the monitored rules. In this case we terminate the session.

Reviewed By: mpgermano

Differential Revision: D19841921

fbshipit-source-id: 7fb2b2f558d23097f91c65d98dec4d17f24051a0