symphony/circleci: bump terraform version
Summary:
In addition change resource count calculation as I was getting the following errors with new version:
```
Error: Invalid count argument

  on orc8r.tf line 176, in resource "aws_iam_role_policy" "orc8r_admin":
 176:   count  = length(aws_iam_role.orc8r_admin)

The "count" value depends on resource attributes that cannot be determined
until apply, so Terraform cannot predict how many instances will be created.
To work around this, use the -target argument to first apply only the
resources that the count depends on.

Error: Invalid count argument

  on orc8r.tf line 183, in resource "aws_iam_role_policy_attachment" "orc8r_admin_read_only":
 183:   count      = length(aws_iam_role.orc8r_admin)

The "count" value depends on resource attributes that cannot be determined
until apply, so Terraform cannot predict how many instances will be created.
To work around this, use the -target argument to first apply only the
resources that the count depends on.
```

Reviewed By: a8m

Differential Revision: D20053478

fbshipit-source-id: 14bb7721f3d1855cdc1f295feab9e79b50a18300