Improve MockRedisPipeline to store command results
Summary:
Recently released version 0.8.0 of `redis-collections` uses a Redis pipe object
more than it had previously. Our mock implementation to this point served as a no-op.
This was breaking unit tests. This diff improves the functionality of the mock pipeline
by storing the result's of the commands that are run.

Reviewed By: themarwhal

Differential Revision: D20076768

fbshipit-source-id: 9611ef787c5c7d68fb3dce6e6fdc84e1ae8b7064