Fixing retrieval of spgw bearer context info on actv/deactv bearer handlers (#1247)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1247

On PGW handling of NW initiated activate / deactivate bearer requests, with updates of usages of SPGW IMSI => Tunnel ID map,
as there can be multiple entries with the same IMSI for multiple PDN connections,
the retrieval of SPGW UE context info state by IMSI is not sufficient as it will always return the last inserted entry,
which causes a bug. This diff fixes it.

- On `pgw_handle_nw_initiated_bearer_deactv_req`, it now traverses entire map to match by IMSI + EPS Bearer LBI.

Reviewed By: ulaskozat

Differential Revision: D20100843

fbshipit-source-id: 54eadbf49f4c10cd35a66ad60c0046135c343d4e