Change default value for extra_quota_margin to 1024
Summary:
There's a sessiond config called `extra_quota_margin` that affects when session's get terminated.
The basic idea is that even if we've used up the charging quota and it is not the final grant, if it's excessively above the quota limit, we will terminate the service.
This was added to tackle the case where the UpdateRequest to the reporter (FeG, or PolicyDB) failed. If the GRPC call fails permanently, we would eventually want to terminate the service.

The issue here was the default value in the sessiond config was set to 0. Which meant that if pipelined reports > 100% of the charging quota grant, we would always terminate the session.

I'm not sure if I like this solution because you can always find a data rate at which the session will always get terminated even if there's no failure. Maybe we can keep track of the number of failed grpc calls and terminate at a threshold.

But in the meantime, I'll change the sessiond logic to take the max of (config value, default=1024bytes).

Differential Revision: D20110872

fbshipit-source-id: cfa18d974e95b8d44a74edf2def5ffc7bc918255