Enforce Pipelined app ordering
Summary:
Its necessary to enforce a logical order for some pipelined apps. This add a priority ordering that should resolve that.
Also bring out enforcement stats into a separate dedicated table, this is done because it needs to be resilient and having all the rules in a temporary on demand scratch table just doesn't work.

100 test file snapshots coming in next diff

Reviewed By: mpgermano

Differential Revision: D20163428

fbshipit-source-id: db3b34e67ded96d786b71ca49b5e98026c56ab20