Nuke stale pipelined check_quota arp entries
Summary:
on restarts we don't want to have any arp entries already set for the fake check quota IPs.

Unfortunately we can't use this from inside docker so we can't remove arp with `arp -d`. Hopefully subnet of /16 is enough + arp timeout.

I'll look into fixing the behavior with the wrong arp in the near future.

Reviewed By: ardzoht

Differential Revision: D20201718

fbshipit-source-id: 9a527b4b948ee88cc6bc13bddef3838faccedd41