FIX ESLINT ON VSCODE
Summary:
eslint wasn't working on VS Code @ FB. This is because eslint is only configured to run on `javascript`, `javascriptreact`, and a few other Language IDs in VSCode. The files we use are almost exclusively `Flow` files.

Adding this to the list of languages to run eslint on actually gets eslint to run.

Reviewed By: tcirstea

Differential Revision: D20239032

fbshipit-source-id: d1065d91270eacf0e65d2cc472ff2fc8d6ff964e